+++ isitfit/tests/test_utils.py
@@ -2,7 +2,7 @@
 # https://github.com/pytest-dev/pytest-mock
 def test_ping_matomo(mocker):
   from ..utils import ping_matomo
-  def mockreturn(url, json): return "foo"
+  def mockreturn(url, json, timeout): return "foo"
   mocked_post = mocker.patch('isitfit.utils.requests.post', side_effect=mockreturn)
   ping_matomo("/test")